refactor(settings): extract SnappingEffects + TilingAppearance controllers
Two thin bounds facades bundled together — same pattern as
SnappingZoneSelectorController, both inline-header-only with minimal
.cpp shims for AUTOMOC placement.

SnappingEffectsController: shaderFrameRate bounds, audioSpectrumBarCount
bounds, cavaAvailable.

TilingAppearanceController: autotileBorderWidth/Radius bounds.
