+++ src/settings/snappingeffectscontroller.cpp
@@ -1,0 +1,4 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#include "snappingeffectscontroller.h"
